// frontend/Types/dataTypes.ts
export interface achievementType {
  id: number;
  logo: any;
  title: string;
  disc: string;
}